Skin-Cancer-ML/requirement.txt
flask
numpy
pandas
tensorflow
scikit-learn
pillow
openpyxl
